requirements.txt
torch
transformers	
numpy
streamlit
pillow
plotly
lime
